Add files via upload
The model is built on data downloaded from internet. The data has been analyzed and techniques are implemented as per the identified scenarios in the data.
In Part 1- The target variable is highly imbalanced. Here approaches and feature engineering are done with traditional approach.
In Part 2 - The data is balanced, for feature engineering bins for the features are created and cross checked with hypothesis testing then one-hot encoding is done
The final report contains the approach for data collection, cleaning, feature engineering and the solution for business problem.